% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addLinkToEWASDataHub}
\alias{addLinkToEWASDataHub}
\title{addLinkToEWASDataHub
adds links to EWASDataHub to a data.frame as separate column}
\usage{
addLinkToEWASDataHub(df, baseURL)
}
\arguments{
\item{df}{data.frame to which links should be added}

\item{baseURL}{string describing link}
}
\value{
data.frame
}
\description{
addLinkToEWASDataHub
adds links to EWASDataHub to a data.frame as separate column
}
