% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addLinkToEWASDataHubToHeader}
\alias{addLinkToEWASDataHubToHeader}
\title{addLinkToEWASDataHubToHeader
adds links to EWAS data hub to a data.frame into first line}
\usage{
addLinkToEWASDataHubToHeader(df, baseURL)
}
\arguments{
\item{df}{data.frame to which links should be added}
}
\value{
data.frame
}
\description{
addLinkToEWASDataHubToHeader
adds links to EWAS data hub to a data.frame into first line
}
