% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addLinkToMRCEWASCatalog}
\alias{addLinkToMRCEWASCatalog}
\title{addLinkToMRCEWASCatalog
adds links to MRC EWAS catalog to a data.frame as separate column}
\usage{
addLinkToMRCEWASCatalog(df, baseURL)
}
\arguments{
\item{df}{data.frame to which links should be added}

\item{baseURL}{string describing link}
}
\value{
data.frame
}
\description{
addLinkToMRCEWASCatalog
adds links to MRC EWAS catalog to a data.frame as separate column
}
