% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addLinkToMRCEWASCatalogToHeader}
\alias{addLinkToMRCEWASCatalogToHeader}
\title{addLinkToMRCEWASCatalogToHeader
adds links to MRC EWAS catalog to a data.frame into first line}
\usage{
addLinkToMRCEWASCatalogToHeader(df, baseURL)
}
\arguments{
\item{df}{data.frame to which links should be added}
}
\value{
data.frame
}
\description{
addLinkToMRCEWASCatalogToHeader
adds links to MRC EWAS catalog to a data.frame into first line
}
