% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.onAttach}
\alias{.onAttach}
\title{very first function}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\value{
nothing
}
\description{
very first function during package load
}
\keyword{internal}
