% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{getDMPNearRange}
\alias{getDMPNearRange}
\title{gets a data structure with the near range of a DMR}
\usage{
getDMPNearRange(DMP, DMPNearRangeData, traitVar, mergeAttribut, range)
}
\arguments{
\item{DMP}{a data.frame containing row.names(ID), IDs (ID_Kind), sex, trait and cgs around probe}

\item{DMPNearRangeData}{a data.frame containing the CpGs around a selected CpG}

\item{traitVar}{a data.frame containing the real world measurements for a certain trait}

\item{mergeAttribut}{the attribut from traitVar, that is used for merging with DMPNearRangeData}

\item{range}{trait original methylation data as data.frame: probeID, BETA, SE, P_VAL, FDR, DeltaMeth, N}
}
\value{
data.frame
}
\description{
gets a data structure with the near range of a DMR
}
