% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{getDMPNearRangeprobeID}
\alias{getDMPNearRangeprobeID}
\title{gets a data frame with the near range probeIDs of a DMR}
\usage{
getDMPNearRangeprobeID(annotation, DMP, range)
}
\arguments{
\item{annotation}{annotation data (pre loaded); the annotation from meffil}

\item{DMP}{CpG to find the range around}

\item{range}{number of CpG to the left and to the right to include to range (truncated at chromosome end)}
}
\value{
data.frame
}
\description{
gets a data structure with the near range of a DMR
}
