% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getResultDataSingleTrait}
\alias{getResultDataSingleTrait}
\title{getResultDataSingleTrait
gets back the currently selected trait}
\usage{
getResultDataSingleTrait(
  globalVariables,
  sessionVariables,
  onlySignificant = FALSE
)
}
\arguments{
\item{globalVariables}{contains all global available Objects}

\item{sessionVariables}{contains all session objects}
}
\value{
data.frame
}
\description{
getResultDataSingleTrait
gets back the currently selected trait
}
