% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getTraitsDFLong}
\alias{getTraitsDFLong}
\title{getTraitsDFLong
gets back the currently selected trait together with gender information and traitnames were replaced with filename compatible characters}
\usage{
getTraitsDFLong(globalVariables)
}
\arguments{
\item{globalVariables}{contains all global available Objects}
}
\value{
data.frame
}
\description{
getTraitsDFLong
gets back the currently selected trait together with gender information and traitnames were replaced with filename compatible characters
}
