% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadObjects}
\alias{loadObjects}
\title{loadObjects
loads globally needed objects (methylation matrix with beta values, annotation, etc.)}
\usage{
loadObjects(globalVariables)
}
\arguments{
\item{globalVariables}{contains all global available Objects}
}
\value{
globalVariables
}
\description{
loadObjects
loads globally needed objects (methylation matrix with beta values, annotation, etc.)
}
