% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadResultFile}
\alias{loadResultFile}
\title{loadResultFile}
\usage{
loadResultFile(globalVariables, sessionVariables)
}
\arguments{
\item{globalVariables}{globalVariables loads files with globally available information (beta, trait)}

\item{sessionVariables}{sessionVariables}
}
\value{
data.frame with regression results
}
\description{
loadResultFile
}
