% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{plotlyPcPForDMP}
\alias{plotlyPcPForDMP}
\title{Plots the near range of a DMR}
\usage{
plotlyPcPForDMP(
  DMPNearRange,
  probe,
  resultDataSingleTrait,
  annotation,
  selection,
  shortlabel = TRUE
)
}
\arguments{
\item{DMPNearRange}{a data.frame containing row.names(ID), IDs (ID_Kind), sex, trait and cgs around probe}

\item{probe}{probe (in the center), that the parallel coordinate plot should show; cg number}

\item{resultDataSingleTrait}{trait original methylation data as data.frame: probeID, BETA, SE, P_VAL, FDR, DeltaMeth, N}

\item{annotation}{annotation data (pre loaded); the annotation from meffil}

\item{selection}{a list of cg numbers to mark within the methylation profile plot}

\item{shortlabel}{if TRUE only the cg number is plotted into the methylation profile plot}
}
\value{
plotly object
}
\description{
plots the near range of a DMR in a parallel coordinate plot and gives a plotly object back
}
