% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{plotlyPcPForDMPSessionVariables}
\alias{plotlyPcPForDMPSessionVariables}
\title{Plots the near range of a DMR}
\usage{
plotlyPcPForDMPSessionVariables(
  globalVariables,
  sessionVariables,
  DMPNearRange
)
}
\description{
plots the near range of a DMR and gives a plotly object back using session variables
}
