% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{plotlyViolinForDMP}
\alias{plotlyViolinForDMP}
\title{Plots the violin plot for both sexes of a near range of a DMR}
\usage{
plotlyViolinForDMP(sexFemaleValue, sexMaleValue, DMPNearRange)
}
\description{
plots the violin plot for both sexes of a near range of a DMR and gives a plotly object back
}
