% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{reducedAnnotation}
\alias{reducedAnnotation}
\title{reducedAnnotation
expects a data.frame with annotation for Illuminas 450k array like provided by meffil package
removes chromosome X and Y from annotation data.frame and sorts it by chromosome number and position
removes attributes type, target and meth.dye from data.frame}
\usage{
reducedAnnotation(a)
}
\arguments{
\item{a}{data.frame to which links should be added}
}
\value{
data.frame
}
\description{
reducedAnnotation
expects a data.frame with annotation for Illuminas 450k array like provided by meffil package
removes chromosome X and Y from annotation data.frame and sorts it by chromosome number and position
removes attributes type, target and meth.dye from data.frame
}
\keyword{internal}
