% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeMultiModelCpGFromBeta}
\alias{removeMultiModelCpGFromBeta}
\title{removeMultiModelCpGFromBeta
removes multimodal CpG from data.frame}
\usage{
removeMultiModelCpGFromBeta(df, multiModList)
}
\arguments{
\item{multiModList}{list with multimodal CpG}

\item{data.frame}{to which links should be added}
}
\value{
data.frame
}
\description{
removeMultiModelCpGFromBeta
removes multimodal CpG from data.frame
}
