% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{resultDataSingleScenarioWithAnnotation}
\alias{resultDataSingleScenarioWithAnnotation}
\title{resultDataSingleScenarioWithAnnotation
expects annotation and data.frame
removes long stromgs from gene.symbol
removes attributes accession, region, cpg.island.name, relation.to.island and snp.exclude from data.frame}
\usage{
resultDataSingleScenarioWithAnnotation(annotation, df)
}
\arguments{
\item{annotation}{data.frame to which links should be added}

\item{df}{data.frame to which links should be added}
}
\value{
data.frame
}
\description{
resultDataSingleScenarioWithAnnotation
expects annotation and data.frame
removes long stromgs from gene.symbol
removes attributes accession, region, cpg.island.name, relation.to.island and snp.exclude from data.frame
}
