% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{winsorize}
\alias{winsorize}
\title{winsorize
performs winsorizing}
\usage{
winsorize(traitDF, trim, startVar, endVar)
}
\arguments{
\item{traitDF}{data.frame to be used}

\item{trim}{value to be used for winsorizing}

\item{startVar}{variable to start with}

\item{endVar}{variable to end at}
}
\value{
data.frame with winsorized variables
}
\description{
winsorize
performs winsorizing
}
\examples{
EpiVisR::winsorize(df, 0.05, 10, 20)
}
