% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.onAttach}
\alias{.onAttach}
\title{very first function}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\value{
nothing
}
\description{
very first function during package load
}
<<<<<<< HEAD
\keyword{.onAttach()}
=======
>>>>>>> 293e544a2d61b0de48d77a832b4da81bba457b80
\keyword{internal}
