% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{getDMPNearRange}
\alias{getDMPNearRange}
\title{gets a data structure with the near range of a DMR}
\usage{
getDMPNearRange(DMP, DMPNearRangeData, traitVar, mergeAttribut, range)
}
\arguments{
\item{DMP}{#tbc()a data.frame containing row.names(ID), IDs (ID_Kind), sex, trait and cgs around probe}

\item{range}{trait original methylation data as data.frame: probeID, BETA, SE, P_VAL, FDR, DeltaMeth, N}

\item{probe}{#tbc  probe (in the center), that the parallel coordinate plot should show; cg number}
}
\value{
data.frame
}
\description{
gets a data structure with the near range of a DMR
}
