% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadResultFile}
\alias{loadResultFile}
\title{loadResultFile}
\usage{
loadResultFile(sessionVariables, numberResults = 100)
}
\arguments{
\item{sessionVariables}{sessionVariables loads files with globally available information (beta, trait)}

\item{numberResults}{limits the number of rows to load for faster debugging}
}
\value{
data.frame with regression results
}
\description{
loadResultFile
}
