% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAmProfile.R
\name{plotDNAmProfile_SERVER}
\alias{plotDNAmProfile_SERVER}
\title{Plots the near range of a DMR}
\usage{
plotDNAmProfile_SERVER(id, globalVariables, sessionVariables)
}
\arguments{
\item{id}{id to identify against UI}

\item{globalVariables}{package global variables}

\item{sessionVariables}{package session variables}
}
\description{
plots the near range of a DMR defined inside session variables
}
